refactor : Refactor integrate RacingGame, RacingSetUp
