+++ src/main/java/racingcar/Application.java
@@ -1,16 +1,17 @@
 package racingcar;
 
 import racingcar.controller.RacingGame;
-import racingcar.controller.RacingSetUp;
 import racingcar.domain.MoveStrategy;
 import racingcar.domain.RandomMoveStrategy;
+import racingcar.view.InputView;
 
 public class Application {
     public static void main(String[] args) {
-        RacingSetUp racingSetUp = new RacingSetUp();
+        String carNames = InputView.inputCarNames();
+        String tryCount = InputView.inputTryCount();
         MoveStrategy moveStrategy = new RandomMoveStrategy();
 
-        RacingGame racingGame = new RacingGame(racingSetUp, moveStrategy);
+        RacingGame racingGame = new RacingGame(carNames, tryCount, moveStrategy);
         racingGame.start();
     }
 }
